enviando alterações de responsividade
