+++ pokedex/src/pages/PokeDetailsPage/styles.js
@@ -1,0 +1,14 @@
+import styled from "styled-components"
+
+export const StyleCardDetails = styled.div`
+text-align: center;
+border: 5px solid  #FECB05;
+background-color: #c9cdde;
+padding-bottom: 15px;
+font-family: Arial, Helvetica, sans-serif;
+color: #1F356B;
+
+img {
+    width: 20%;
+}
+`;
